+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Changed
 
 ### Breaking changes
+- Removing the `tempValueHolder` logic from the `enchanted-textfield`
 
 ## 3.0.0